SessionSetup: Don't check cmd when detecting MORE_PROCESSING
This should only happen if the command was the initial phase
of NTLM auth so we don't need to check it here.
And if  we get this when cmd==3 i.e. as a response to an AUTH
packet then authentication is busted and we will just
try again until we get a har failure.

This saves 18 bytes when compiled to x86_64

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>